Table: Refactor TableNG nested implementation (flat/nested split)
Squashed refactor work: splits the monolithic TableNG into a flat/nested
dispatcher with shared TableDataGrid wrapper and render-hooks, fixes nested
row expansion state across refreshes, adds auto-expand for nested rows, and
adds the proto-getter row parser. Builds on the table.refactorNested flag.
