+++ packages/grafana-ui/src/components/Table/TableNG/TableDataGrid.test.tsx
@@ -1,0 +1,69 @@
+import { render, screen } from '@testing-library/react';
+import { createRef } from 'react';
+
+import { type DataGridHandle } from '@grafana/react-data-grid';
+
+import { TableDataGrid, type TableDataGridProps } from './TableDataGrid';
+
+function makeProps(overrides: Partial<TableDataGridProps> = {}): TableDataGridProps {
+  return {
+    role: 'grid',
+    gridRef: createRef<DataGridHandle>(),
+    columns: [],
+    rows: [],
+    renderers: {
+      renderRow: jest.fn(),
+      renderCell: jest.fn(),
+    },
+    onCellClick: jest.fn(),
+    onCellKeyDown: jest.fn(),
+    sortColumns: [],
+    setSortColumns: jest.fn(),
+    rowHeight: 36,
+    hasFooter: false,
+    footerHeight: 0,
+    noHeader: false,
+    headerHeight: 36,
+    enablePagination: false,
+    numRows: 0,
+    page: 0,
+    setPage: jest.fn(),
+    numPages: 1,
+    pageRangeStart: 0,
+    pageRangeEnd: 0,
+    smallPagination: false,
+    sortedRows: [],
+    onTooltipClose: jest.fn(),
+    onInspectCellDismiss: jest.fn(),
+    ...overrides,
+  };
+}
+
+describe('TableDataGrid', () => {
+  let origResizeObserver = global.ResizeObserver;
+
+  beforeEach(() => {
+    origResizeObserver = global.ResizeObserver;
+    global.ResizeObserver = class ResizeObserver {
+      observe() {}
+      unobserve() {}
+      disconnect() {}
+    };
+  });
+
+  afterEach(() => {
+    global.ResizeObserver = origResizeObserver;
+  });
+
+  describe('DataGrid prop pass-through', () => {
+    it('forwards data-testid to the underlying grid element', () => {
+      render(<TableDataGrid {...makeProps({ 'data-testid': 'my-custom-grid' })} />);
+      expect(screen.getByTestId('my-custom-grid')).toBeInTheDocument();
+    });
+
+    it('forwards aria-label to the underlying grid element', () => {
+      render(<TableDataGrid {...makeProps({ 'aria-label': 'accessible grid' })} />);
+      expect(screen.getByRole('grid', { name: 'accessible grid' })).toBeInTheDocument();
+    });
+  });
+});
