+++ packages/grafana-ui/src/components/Table/TableNG/TableFlat.tsx
@@ -1,0 +1,355 @@
+import memoize from 'micro-memoize';
+import { useCallback, useMemo, useRef, useState } from 'react';
+
+import { type Field } from '@grafana/data';
+import { type DataGridHandle, type DataGridProps } from '@grafana/react-data-grid';
+
+import { useTheme2 } from '../../../themes/ThemeContext';
+import { getTextColorForBackground as _getTextColorForBackground } from '../../../utils/colors';
+import { usePanelContext } from '../../PanelChrome';
+import { type DataLinksActionsTooltipState } from '../cellUtils';
+
+import { TableDataGrid } from './TableDataGrid';
+import { TABLE } from './constants';
+import {
+  useColumnResize,
+  useColWidths,
+  useFlatRowHeight,
+  useFilteredRows,
+  useHeaderHeight,
+  useManagedSort,
+  usePaginatedRows,
+  useScrollbarWidth,
+  useSortedRows,
+} from './hooks';
+import { type ColumnBuildConfig, useColumnBuilderFromFields, useDataGridRows } from './render-hooks';
+import {
+  type CellRootRenderer,
+  type InspectCellProps,
+  type TableColumn,
+  type TableNGProps,
+  type TableRow,
+  type TableSummaryRow,
+} from './types';
+import {
+  calculateFooterHeight,
+  compileFrameToRecordsV1,
+  compileFrameToRecordsV2,
+  createTypographyContext,
+  extractPixelValue,
+  getApplyToRowBgFn,
+  getCellColorInlineStylesFactory,
+  getCellLinks,
+  getDefaultRowHeight,
+  getVisibleFields,
+} from './utils';
+
+type OnCellClick = NonNullable<DataGridProps<TableRow, TableSummaryRow>['onCellClick']>;
+
+// Flat tables have no depth-1 rows, so expandedRows is never consulted.
+// Stable references avoid invalidating useDataGridRows' memo on every render.
+const EMPTY_EXPANDED_ROWS: Set<string> = new Set();
+const NOOP_STABLE_KEY = () => '';
+
+export function TableFlat(props: TableNGProps) {
+  const {
+    cellHeight,
+    data,
+    disableKeyboardEvents,
+    disableSanitizeHtml,
+    enablePagination = false,
+    enableSharedCrosshair = false,
+    enableVirtualization,
+    frozenColumns: _frozenColumns = 0,
+    getActions = () => [],
+    height,
+    maxRowHeight: _maxRowHeight,
+    noHeader,
+    noValue,
+    onCellFilterAdded,
+    onColumnResize,
+    onSortByChange,
+    protoParserEnabled,
+    showTypeIcons,
+    structureRev,
+    timeRange,
+    transparent,
+    width,
+    initialRowIndex,
+    sortBy,
+    sortByBehavior = 'initial',
+  } = props;
+
+  const theme = useTheme2();
+  const panelContext = usePanelContext();
+  const userCanExecuteActions = useMemo(() => panelContext.canExecuteActions?.() ?? false, [panelContext]);
+
+  const getCellActions = useCallback(
+    (field: Field, rowIdx: number) => {
+      if (!userCanExecuteActions) {
+        return [];
+      }
+      return getActions(data, field, rowIdx);
+    },
+    [getActions, data, userCanExecuteActions]
+  );
+
+  const visibleFields = useMemo(() => getVisibleFields(data.fields), [data.fields]);
+  const hasHeader = !noHeader;
+  const hasFooter = useMemo(
+    () => visibleFields.some((field) => Boolean(field.config.custom?.footer?.reducers?.length)),
+    [visibleFields]
+  );
+  const footerHeight = useMemo(
+    () => (hasFooter ? calculateFooterHeight(visibleFields) : 0),
+    [hasFooter, visibleFields]
+  );
+
+  const resizeHandler = useColumnResize(onColumnResize);
+
+  const frameToRecords = useMemo(
+    () => (protoParserEnabled ? compileFrameToRecordsV2(data, undefined) : compileFrameToRecordsV1(data, undefined)),
+    [data, protoParserEnabled]
+  );
+  const rows = useMemo(() => frameToRecords(data), [frameToRecords, data]);
+
+  const { rows: filteredRows, filter, setFilter, filterResult } = useFilteredRows(rows, data.fields);
+  const {
+    rows: sortedRows,
+    sortColumns,
+    setSortColumns,
+  } = useSortedRows(filteredRows, data.fields, [], { initialSortBy: sortBy });
+
+  useManagedSort({ sortByBehavior, setSortColumns, sortBy });
+
+  const [inspectCell, setInspectCell] = useState<InspectCellProps | null>(null);
+  const [tooltipState, setTooltipState] = useState<DataLinksActionsTooltipState>();
+  const onCellClick: OnCellClick = useCallback(
+    ({ column, row }, ev) => {
+      // eslint-disable-next-line @typescript-eslint/consistent-type-assertions
+      const field = (column as unknown as TableColumn).field;
+
+      if (ev.target instanceof HTMLElement && ev.target.closest('a[aria-haspopup], .rdg-cell')?.matches('a')) {
+        const rowIdx = row.__index;
+        setTooltipState({
+          coords: { clientX: ev.clientX, clientY: ev.clientY },
+          links: getCellLinks(field, rowIdx),
+          actions: getCellActions(field, rowIdx),
+        });
+        ev.preventGridDefault();
+      }
+    },
+    [getCellActions]
+  );
+
+  const gridRef = useRef<DataGridHandle>(null);
+  const scrollbarWidth = useScrollbarWidth(gridRef, height);
+  const availableWidth = useMemo(() => width - scrollbarWidth, [width, scrollbarWidth]);
+
+  const getCellColorInlineStyles = useMemo(() => getCellColorInlineStylesFactory(theme), [theme]);
+  const applyToRowBgFn = useMemo(
+    () => getApplyToRowBgFn(data.fields, getCellColorInlineStyles) ?? undefined,
+    [data.fields, getCellColorInlineStyles]
+  );
+  const getTextColorForBackground = useMemo(() => memoize(_getTextColorForBackground, { maxSize: 1000 }), []);
+
+  const typographyCtx = useMemo(
+    () =>
+      createTypographyContext(
+        theme.typography.fontSize,
+        theme.typography.fontFamily,
+        extractPixelValue(theme.typography.body.letterSpacing!) * theme.typography.fontSize
+      ),
+    [theme]
+  );
+
+  const frozenColumns = _frozenColumns;
+
+  // When a width override is removed from field config, the configured-width count drops. That
+  // change to field.config.custom.width is a mutation on the existing field objects, so it doesn't
+  // re-trigger memoization on its own. We detect the drop here and pass a fresh reset key to force
+  // recomputation and clear react-data-grid's internal column widths so columns re-flow to auto.
+  const configuredWidthCount = visibleFields.reduce(
+    (count, field) => count + (field.config.custom?.width != null ? 1 : 0),
+    0
+  );
+  const prevConfiguredWidthCount = useRef(configuredWidthCount);
+  const widthConfigResetKey = configuredWidthCount < prevConfiguredWidthCount.current ? Symbol() : undefined;
+  const resetColumnWidths = widthConfigResetKey != null ? new Map() : undefined;
+
+  prevConfiguredWidthCount.current = configuredWidthCount;
+
+  const [widths, numFrozenColsFullyInView] = useColWidths(
+    visibleFields,
+    availableWidth,
+    frozenColumns,
+    widthConfigResetKey
+  );
+
+  const headerHeight = useHeaderHeight({
+    columnWidths: widths,
+    fields: visibleFields,
+    enabled: hasHeader,
+    sortColumns,
+    showTypeIcons: showTypeIcons ?? false,
+    typographyCtx,
+  });
+  const maxRowHeight = _maxRowHeight != null ? Math.max(TABLE.LINE_HEIGHT, _maxRowHeight) : undefined;
+
+  const defaultRowHeight = useMemo(
+    () => getDefaultRowHeight(theme, visibleFields, cellHeight),
+    [theme, visibleFields, cellHeight]
+  );
+
+  const rowHeight = useFlatRowHeight({
+    columnWidths: widths,
+    fields: visibleFields,
+    defaultHeight: defaultRowHeight,
+    typographyCtx,
+    maxHeight: maxRowHeight,
+  });
+
+  const {
+    rows: paginatedRows,
+    page,
+    setPage,
+    numPages,
+    numRows,
+    pageRangeStart,
+    pageRangeEnd,
+    smallPagination,
+  } = usePaginatedRows(sortedRows, {
+    enabled: enablePagination,
+    width: availableWidth,
+    height,
+    footerHeight,
+    headerHeight: hasHeader ? headerHeight : 0,
+    rowHeight,
+  });
+
+  const rowHeightFn = useMemo((): ((row: TableRow) => number) => {
+    if (typeof rowHeight === 'function') {
+      // eslint-disable-next-line @typescript-eslint/consistent-type-assertions
+      return rowHeight as unknown as (row: TableRow) => number;
+    }
+    if (typeof rowHeight === 'string') {
+      return () => TABLE.MAX_CELL_HEIGHT;
+    }
+    return () => rowHeight;
+  }, [rowHeight]);
+
+  const renderRow = useDataGridRows(
+    data.fields,
+    panelContext,
+    EMPTY_EXPANDED_ROWS,
+    enableSharedCrosshair,
+    NOOP_STABLE_KEY
+  );
+
+  const columnBuildConfig = useMemo(
+    (): ColumnBuildConfig => ({
+      theme,
+      applyToRowBgFn,
+      getCellColorInlineStyles,
+      getTextColorForBackground,
+      rowHeight,
+      rowHeightFn,
+      filter,
+      setFilter,
+      setInspectCell,
+      gridRef,
+      getCellActions,
+      onCellFilterAdded,
+      frozenColumns,
+      numFrozenColsFullyInView,
+      maxRowHeight,
+      disableKeyboardEvents,
+      disableSanitizeHtml,
+      showTypeIcons,
+      timeRange,
+    }),
+    [
+      theme,
+      applyToRowBgFn,
+      getCellColorInlineStyles,
+      getTextColorForBackground,
+      rowHeight,
+      rowHeightFn,
+      filter,
+      getCellActions,
+      onCellFilterAdded,
+      frozenColumns,
+      numFrozenColsFullyInView,
+      maxRowHeight,
+      disableKeyboardEvents,
+      disableSanitizeHtml,
+      setFilter,
+      showTypeIcons,
+      timeRange,
+    ]
+  );
+
+  const fromFields = useColumnBuilderFromFields(filterResult, columnBuildConfig);
+
+  const { columns, cellRootRenderers } = useMemo(
+    () => fromFields(visibleFields, widths, data, rows, sortedRows),
+    [fromFields, visibleFields, widths, data, rows, sortedRows]
+  );
+
+  // invalidate columns on every structureRev change to support width editing in fieldConfig.
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+  const structureRevColumns = useMemo(() => columns, [columns, structureRev]);
+  const renderCellRoot: CellRootRenderer = useCallback(
+    (key, cellProps) => cellRootRenderers[cellProps.column.key](key, cellProps),
+    [cellRootRenderers]
+  );
+
+  return (
+    <TableDataGrid
+      role="grid"
+      gridRef={gridRef}
+      columns={structureRevColumns}
+      rows={paginatedRows}
+      noValue={noValue}
+      renderers={{ renderRow, renderCell: renderCellRoot }}
+      columnWidths={resetColumnWidths}
+      onColumnWidthsChange={resetColumnWidths != null ? () => {} : undefined}
+      onColumnResize={resizeHandler}
+      onCellClick={onCellClick}
+      onCellKeyDown={({ column, row }, event) => {
+        if (column.key === columns[0].key && row.__index === 0 && event.shiftKey && event.key === 'Tab') {
+          event.preventGridDefault();
+          gridRef.current?.selectCell({ rowIdx: -1, idx: columns.length - 1 });
+          return;
+        }
+        if (disableKeyboardEvents) {
+          event.preventGridDefault();
+        }
+      }}
+      sortColumns={sortColumns}
+      setSortColumns={setSortColumns}
+      onSortByChange={onSortByChange}
+      rowHeight={rowHeight}
+      enableVirtualization={enableVirtualization}
+      hasFooter={hasFooter}
+      footerHeight={footerHeight}
+      noHeader={!!noHeader}
+      headerHeight={headerHeight}
+      transparent={transparent}
+      initialRowIndex={initialRowIndex}
+      sortedRows={sortedRows}
+      enablePagination={enablePagination}
+      numRows={numRows}
+      page={page}
+      setPage={setPage}
+      numPages={numPages}
+      pageRangeStart={pageRangeStart}
+      pageRangeEnd={pageRangeEnd}
+      smallPagination={smallPagination}
+      tooltipState={tooltipState}
+      onTooltipClose={() => setTooltipState(undefined)}
+      inspectCell={inspectCell}
+      onInspectCellDismiss={() => setInspectCell(null)}
+    />
+  );
+}
