+++ packages/grafana-ui/src/components/Table/TableNG/TableNG.tsx
@@ -1,1142 +1,29 @@
-import '@grafana/react-data-grid/lib/styles.css';
+import { Suspense, useMemo } from 'react';
 
-import { clsx } from 'clsx';
-import memoize from 'micro-memoize';
-import {
-  type CSSProperties,
-  type JSX,
-  type Key,
-  type ReactNode,
-  Suspense,
-  useCallback,
-  useEffect,
-  useId,
-  useMemo,
-  useRef,
-  useState,
-} from 'react';
+import { FieldType } from '@grafana/data';
 
-import {
-  type DataFrame,
-  DataHoverClearEvent,
-  DataHoverEvent,
-  FALLBACK_COLOR,
-  type Field,
-  FieldType,
-  getDisplayProcessor,
-} from '@grafana/data';
-import { t, Trans } from '@grafana/i18n';
-import {
-  Cell,
-  type CellRendererProps,
-  DataGrid,
-  type DataGridHandle,
-  type DataGridProps,
-  type RenderCellProps,
-  type Renderers,
-  type RenderRowProps,
-  Row,
-  type SortColumn,
-} from '@grafana/react-data-grid';
-import { FieldColorModeId, TableCellTooltipPlacement, type TableFooterOptions } from '@grafana/schema';
-
-import { useStyles2, useTheme2 } from '../../../themes/ThemeContext';
-import { getTextColorForBackground as _getTextColorForBackground } from '../../../utils/colors';
-import { Pagination } from '../../Pagination/Pagination';
-import { type PanelContext, usePanelContext } from '../../PanelChrome';
-import { DataLinksActionsTooltip } from '../DataLinksActionsTooltip';
-import { TableCellInspector, TableCellInspectorMode } from '../TableCellInspector';
-import { type DataLinksActionsTooltipState } from '../cellUtils';
+import { useStyles2 } from '../../../themes/ThemeContext';
 import { hasGeoCell, LazyOpenLayersProvider } from '../geo';
-import { TableCellDisplayMode } from '../types';
 
-import { getCellRenderer, getCellSpecificStyles } from './Cells/renderers';
-import { EmptyTablePlaceholder } from './components/EmptyTablePlaceholder';
-import { HeaderCell } from './components/HeaderCell';
-import { RowExpander } from './components/RowExpander';
-import { SummaryCell } from './components/SummaryCell';
-import { TableCellActions } from './components/TableCellActions';
-import { TableCellTooltip } from './components/TableCellTooltip';
-import { COLUMN, TABLE } from './constants';
-import {
-  useColumnResize,
-  useColWidths,
-  useFilteredRows,
-  useHeaderHeight,
-  useManagedSort,
-  useNestedColWidths,
-  useNestedRows,
-  usePaginatedRows,
-  useRowHeight,
-  useScrollbarWidth,
-  useSortedRows,
-} from './hooks';
-import {
-  getCellActionStyles,
-  getDefaultCellStyles,
-  getGridStyles,
-  getHeaderCellStyles,
-  getLinkStyles,
-  getMaxHeightCellStyles,
-  getTooltipStyles,
-} from './styles';
-import {
-  type CellRootRenderer,
-  type FromFieldsResult,
-  type InspectCellProps,
-  type TableCellStyleOptions,
-  type TableColumn,
-  type TableNGProps,
-  type TableRow,
-  type TableSummaryRow,
-} from './types';
-import {
-  calculateFooterHeight,
-  canFieldBeColorized,
-  compileFrameToRecordsV1,
-  compileFrameToRecordsV2,
-  createTypographyContext,
-  displayJsonValue,
-  extractPixelValue,
-  getAlignment,
-  getApplyToRowBgFn,
-  getCellColorInlineStylesFactory,
-  getCellLinks,
-  getCellOptions,
-  getDefaultRowHeight,
-  getDisplayName,
-  getIsNestedTable,
-  getJustifyContent,
-  getStableRowKey,
-  getSummaryCellTextAlign,
-  getVisibleFields,
-  IS_SAFARI_26,
-  isCellInspectEnabled,
-  parseStyleJson,
-  predicateByName,
-  rowKeyGetter,
-  shouldTextOverflow,
-  shouldTextWrap,
-} from './utils';
-
-const EXPANDED_COLUMN_KEY = 'expanded';
-type OnCellClick = NonNullable<DataGridProps<TableRow, TableSummaryRow>['onCellClick']>;
+import { TableFlat } from './TableFlat';
+import { TableNested } from './TableNested';
+import { getGridStyles, IS_SAFARI_26 } from './styles';
+import { type TableNGProps } from './types';
 
 export function TableNG(props: TableNGProps) {
-  const {
-    cellHeight,
-    data,
-    disableKeyboardEvents,
-    disableSanitizeHtml,
-    enablePagination = false,
-    enableSharedCrosshair = false,
-    enableVirtualization,
-    frozenColumns: _frozenColumns = 0,
-    getActions = () => [],
-    height,
-    maxRowHeight: _maxRowHeight,
-    noHeader,
-    noValue,
-    onCellFilterAdded,
-    onColumnResize,
-    onSortByChange,
-    protoParserEnabled,
-    showTypeIcons,
-    structureRev,
-    timeRange,
-    transparent,
-    width,
-    initialRowIndex,
-    sortBy,
-    sortByBehavior = 'initial',
-  } = props;
-  const uniqueId = useId();
-  const theme = useTheme2();
-
-  const panelContext = usePanelContext();
-  const userCanExecuteActions = useMemo(() => panelContext.canExecuteActions?.() ?? false, [panelContext]);
-
-  const getCellActions = useCallback(
-    (field: Field, rowIdx: number) => {
-      if (!userCanExecuteActions) {
-        return [];
-      }
-      return getActions(data, field, rowIdx);
-    },
-    [getActions, data, userCanExecuteActions]
-  );
-
-  const visibleFields = useMemo(() => getVisibleFields(data.fields), [data.fields]);
-  const hasHeader = !noHeader;
-  const hasFooter = useMemo(
-    () => visibleFields.some((field) => Boolean(field.config.custom?.footer?.reducers?.length)),
-    [visibleFields]
-  );
-  const footerHeight = useMemo(
-    () => (hasFooter ? calculateFooterHeight(visibleFields) : 0),
-    [hasFooter, visibleFields]
-  );
-
-  const resizeHandler = useColumnResize(onColumnResize);
-  const nestedResizeHandler = useColumnResize(onColumnResize, 'nested');
-
-  const hasNestedFrames = useMemo(() => getIsNestedTable(data.fields), [data]);
+  const { data, transparent } = props;
+  // useStyles2 is always called (Rules of Hooks) even though only safariWrapper is consumed here.
+  // Child components call useStyles2 independently with the correct showPagination value.
+  const styles = useStyles2(getGridStyles, false, transparent);
+  const nestedDataField = useMemo(() => data.fields.find((f) => f.type === FieldType.nestedFrames), [data.fields]);
   const tableHasGeoCell = useMemo(() => hasGeoCell(data), [data]);
-  const nestedFramesFieldName = useMemo(() => {
-    if (!hasNestedFrames) {
-      return;
-    }
-    const firstNestedField = data.fields.find((f) => f.type === FieldType.nestedFrames);
-    if (!firstNestedField) {
-      return;
-    }
-    return getDisplayName(firstNestedField);
-  }, [data, hasNestedFrames]);
-  const frameToRecords = useMemo(
-    () =>
-      protoParserEnabled
-        ? compileFrameToRecordsV2(data, nestedFramesFieldName)
-        : compileFrameToRecordsV1(data, nestedFramesFieldName),
-    [data, nestedFramesFieldName, protoParserEnabled]
+
+  const inner = nestedDataField ? (
+    <TableNested {...props} nestedFramesField={nestedDataField} />
+  ) : (
+    <TableFlat {...props} />
   );
-  const rows = useMemo(() => frameToRecords(data), [frameToRecords, data]);
-
-  const nestedData = useMemo(
-    (): DataFrame[] | undefined =>
-      hasNestedFrames
-        ? data.fields.find((f) => getDisplayName(f) === nestedFramesFieldName)?.values.map((v) => v[0])
-        : undefined,
-    [data, nestedFramesFieldName, hasNestedFrames]
-  );
-
-  // Returns a stable string key for a row based on the groupBy field values stored in the nested
-  // subframe's meta. Falls back to the string index when no stable key is available (non-grouped data).
-  const getStableRowKeyForRowIdx = useCallback(
-    (rowIdx: number): string => getStableRowKey(rowIdx, nestedData?.[rowIdx]),
-    [nestedData]
-  );
-
-  const firstRowNestedData = useMemo(
-    () => (hasNestedFrames && nestedData ? nestedData[0] : undefined),
-    [nestedData, hasNestedFrames]
-  );
-  const nestedFields = useMemo(() => firstRowNestedData?.fields ?? [], [firstRowNestedData]);
-  const nestedVisibleFields = useMemo(() => getVisibleFields(nestedFields), [nestedFields]);
-  const nestedHasFooter = useMemo(
-    () => nestedVisibleFields.some((field) => Boolean(field.config.custom?.footer?.reducers?.length)),
-    [nestedVisibleFields]
-  );
-  const nestedFooterHeight = useMemo(
-    () => (nestedHasFooter ? calculateFooterHeight(nestedVisibleFields) : 0),
-    [nestedHasFooter, nestedVisibleFields]
-  );
-
-  const { rows: filteredRows, filter, setFilter, filterResult } = useFilteredRows(rows, data.fields, hasNestedFrames);
-
-  const {
-    rows: sortedRows,
-    sortColumns,
-    setSortColumns,
-  } = useSortedRows(filteredRows, data.fields, nestedFields, { hasNestedFrames, initialSortBy: sortBy });
-
-  useManagedSort({ sortByBehavior, setSortColumns, sortBy });
-
-  const nestedRows = useNestedRows(rows, nestedData, hasNestedFrames, nestedFramesFieldName, filter, sortColumns);
-
-  const [inspectCell, setInspectCell] = useState<InspectCellProps | null>(null);
-  const [tooltipState, setTooltipState] = useState<DataLinksActionsTooltipState>();
-  const onCellClick: OnCellClick = useCallback(
-    ({ column, row }, ev) => {
-      // we attach field to the column, but it doesn't
-      // eslint-disable-next-line @typescript-eslint/consistent-type-assertions
-      const field = (column as unknown as TableColumn).field;
-
-      // let the click event through for the expander column, since it has its own click handler for expanding/collapsing rows.
-      if (column.key === EXPANDED_COLUMN_KEY) {
-        return;
-      }
-
-      if (
-        ev.target instanceof HTMLElement &&
-        // this walks up the tree to find either a faux link wrapper or the cell root
-        // it then only proceeds if we matched the faux link wrapper
-        ev.target.closest('a[aria-haspopup], .rdg-cell')?.matches('a')
-      ) {
-        const rowIdx = row.__index;
-        setTooltipState({
-          coords: {
-            clientX: ev.clientX,
-            clientY: ev.clientY,
-          },
-          links: getCellLinks(field, rowIdx),
-          actions: getCellActions(field, rowIdx),
-        });
-        ev.preventGridDefault();
-      }
-    },
-    [getCellActions]
-  );
-  const [expandedRows, setExpandedRows] = useState<Set<string>>(() => {
-    if (data.meta?.custom?.expandAllRows) {
-      const nestedField = data.fields.find((f) => f.type === FieldType.nestedFrames);
-      return new Set(Array.from({ length: data.length }, (_, i) => getStableRowKey(i, nestedField?.values[i]?.[0])));
-    }
-    return new Set();
-  });
-  const [selectedRows, setSelectedRows] = useState((): ReadonlySet<string> => new Set());
-
-  // vt scrollbar accounting for column auto-sizing
-  const gridRef = useRef<DataGridHandle>(null);
-  const scrollbarWidth = useScrollbarWidth(gridRef, height);
-  const availableWidth = useMemo(
-    () => (hasNestedFrames ? width - COLUMN.EXPANDER_WIDTH : width) - scrollbarWidth,
-    [width, hasNestedFrames, scrollbarWidth]
-  );
-  const getCellColorInlineStyles = useMemo(() => getCellColorInlineStylesFactory(theme), [theme]);
-  const applyToRowBgFn = useMemo(
-    () => getApplyToRowBgFn(data.fields, getCellColorInlineStyles) ?? undefined,
-    [data.fields, getCellColorInlineStyles]
-  );
-  const getTextColorForBackground = useMemo(() => memoize(_getTextColorForBackground, { maxSize: 1000 }), []);
-
-  const typographyCtx = useMemo(
-    () =>
-      createTypographyContext(
-        theme.typography.fontSize,
-        theme.typography.fontFamily,
-        extractPixelValue(theme.typography.body.letterSpacing!) * theme.typography.fontSize
-      ),
-    [theme]
-  );
-
-  // https://github.com/grafana/grafana/issues/118984: nested tables don't support frozen columns yet.
-  const frozenColumns = useMemo(() => (hasNestedFrames ? 0 : _frozenColumns), [hasNestedFrames, _frozenColumns]);
-  const configuredWidthCount = visibleFields.reduce(
-    (count, field) => count + (field.config.custom?.width != null ? 1 : 0),
-    0
-  );
-  const prevConfiguredWidthCount = useRef(configuredWidthCount);
-  const widthConfigResetKey = configuredWidthCount < prevConfiguredWidthCount.current ? Symbol() : undefined;
-  const resetColumnWidths = widthConfigResetKey != null ? new Map() : undefined;
-
-  prevConfiguredWidthCount.current = configuredWidthCount;
-
-  const [widths, numFrozenColsFullyInView] = useColWidths(
-    visibleFields,
-    availableWidth,
-    frozenColumns,
-    widthConfigResetKey
-  );
-
-  const headerHeight = useHeaderHeight({
-    columnWidths: widths,
-    fields: visibleFields,
-    enabled: hasHeader,
-    sortColumns,
-    showTypeIcons: showTypeIcons ?? false,
-    typographyCtx,
-  });
-  // the minimum max row height we should honor is a single line of text.
-  const maxRowHeight = _maxRowHeight != null ? Math.max(TABLE.LINE_HEIGHT, _maxRowHeight) : undefined;
-  const visibleNestedRowCounts = useMemo(
-    () => nestedRows.map((row, idx) => (expandedRows.has(getStableRowKeyForRowIdx(idx)) ? row.final.length : null)),
-    [nestedRows, expandedRows, getStableRowKeyForRowIdx]
-  );
-
-  const { nestedFieldWidths, nestedColWidths, handleNestedColumnWidthsChange } = useNestedColWidths({
-    nestedVisibleFields,
-    availableWidth,
-    structureRev,
-  });
-
-  const hasNestedHeaders = useMemo(() => firstRowNestedData?.meta?.custom?.noHeader !== true, [firstRowNestedData]);
-  const nestedHeaderHeight = useHeaderHeight({
-    columnWidths: nestedFieldWidths,
-    fields: nestedVisibleFields,
-    enabled: hasNestedHeaders,
-    sortColumns,
-    showTypeIcons: showTypeIcons ?? false,
-    typographyCtx,
-  });
-
-  const defaultRowHeight = useMemo(
-    () => getDefaultRowHeight(theme, visibleFields, cellHeight),
-    [theme, visibleFields, cellHeight]
-  );
-  const defaultNestedRowHeight = useMemo(
-    () => getDefaultRowHeight(theme, nestedVisibleFields, cellHeight),
-    [theme, nestedVisibleFields, cellHeight]
-  );
-
-  const rowHeight = useRowHeight({
-    columnWidths: widths,
-    fields: visibleFields,
-    hasNestedFrames,
-    defaultHeight: defaultRowHeight,
-    defaultNestedHeight: defaultNestedRowHeight,
-    visibleNestedRowCounts,
-    typographyCtx,
-    maxHeight: maxRowHeight,
-    nestedColWidths: nestedFieldWidths,
-    nestedFields: nestedVisibleFields,
-    nestedRows,
-    nestedFooterHeight,
-  });
-
-  const {
-    rows: paginatedRows,
-    page,
-    setPage,
-    numPages,
-    numRows,
-    pageRangeStart,
-    pageRangeEnd,
-    smallPagination,
-  } = usePaginatedRows(sortedRows, {
-    enabled: enablePagination,
-    width: availableWidth,
-    height,
-    footerHeight,
-    headerHeight: hasHeader ? headerHeight : 0,
-    rowHeight,
-    hasNestedFrames,
-  });
-
-  const showPagination = enablePagination && numRows > 0;
-  const styles = useStyles2(getGridStyles, showPagination, transparent);
-
-  const [scrollToIndex, setScrollToIndex] = useState(initialRowIndex);
-  useEffect(() => {
-    if (scrollToIndex !== undefined && sortedRows && gridRef.current?.scrollToCell) {
-      const rowIdx = sortedRows.findIndex((row) => row.__index === scrollToIndex);
-      gridRef.current.scrollToCell({
-        rowIdx,
-      });
-      setScrollToIndex(undefined);
-      setSelectedRows(new Set<string>([rowKeyGetter(sortedRows[rowIdx])]));
-    }
-  }, [scrollToIndex, sortedRows]);
-
-  // normalize the row height into a function which returns a number, so we avoid a bunch of conditionals during rendering.
-  const rowHeightFn = useMemo((): ((row: TableRow) => number) => {
-    if (typeof defaultNestedRowHeight === 'string') {
-      return (row: TableRow) => (expandedRows.has(getStableRowKeyForRowIdx(row.__index)) ? TABLE.MAX_CELL_HEIGHT : 0);
-    }
-    if (typeof rowHeight === 'function') {
-      // this is safe because we only return a (row: TableRow) => string function when defaultNestedRowHeight is a string.
-      // eslint-disable-next-line @typescript-eslint/consistent-type-assertions
-      return rowHeight as unknown as (row: TableRow) => number;
-    }
-    if (typeof rowHeight === 'string') {
-      return () => TABLE.MAX_CELL_HEIGHT;
-    }
-    return () => rowHeight;
-  }, [rowHeight, defaultNestedRowHeight, expandedRows, getStableRowKeyForRowIdx]);
-
-  const renderRow = useMemo(
-    () => renderRowFactory(data.fields, panelContext, expandedRows, enableSharedCrosshair, getStableRowKeyForRowIdx),
-    [data.fields, panelContext, expandedRows, enableSharedCrosshair, getStableRowKeyForRowIdx]
-  );
-
-  const commonDataGridProps = useMemo(
-    () =>
-      ({
-        enableVirtualization: !IS_SAFARI_26 && enableVirtualization !== false && typeof rowHeight !== 'string',
-        defaultColumnOptions: {
-          minWidth: 50,
-          resizable: true,
-          sortable: true,
-          // draggable: true,
-        },
-        onSortColumnsChange: (newSortColumns: SortColumn[]) => {
-          setSortColumns(newSortColumns);
-          onSortByChange?.(
-            newSortColumns.map(({ columnKey, direction }) => ({
-              displayName: columnKey,
-              desc: direction === 'DESC',
-            }))
-          );
-        },
-        sortColumns,
-        rowHeight,
-        bottomSummaryRows: hasFooter ? [{}] : undefined,
-        summaryRowHeight: footerHeight,
-        headerRowClass: styles.headerRow,
-        headerRowHeight: noHeader ? 0 : headerHeight,
-      }) satisfies Partial<DataGridProps<TableRow, TableSummaryRow>>,
-    [
-      enableVirtualization,
-      hasFooter,
-      sortColumns,
-      rowHeight,
-      styles.headerRow,
-      noHeader,
-      setSortColumns,
-      onSortByChange,
-      footerHeight,
-      headerHeight,
-    ]
-  );
-
-  const buildNestedTableExpanderColumn = useCallback(
-    (
-      nestedColumnsMatrix: FromFieldsResult[],
-      hasNestedHeaders: boolean,
-      nestedHeaderHeightPx: number,
-      hasNestedFooter: boolean,
-      nestedFooterHeightPx: number,
-      renderers: Renderers<TableRow, TableSummaryRow>
-    ): TableColumn => ({
-      key: EXPANDED_COLUMN_KEY,
-      sortable: false,
-      resizable: false,
-      name: t('grafana-ui.table.nested-table.expander-column-name', 'Expand nested rows'),
-      field: {
-        name: '',
-        type: FieldType.other,
-        config: {},
-        values: [],
-      },
-      cellClass(row) {
-        if (row.__depth !== 0) {
-          return styles.cellNested;
-        }
-        return;
-      },
-      colSpan(args) {
-        return args.type === 'ROW' && args.row.__depth === 1 ? data.fields.length : 1;
-      },
-      renderCell: ({ row }) => {
-        const rowId = `${uniqueId}-nested-table-${row.__index}`;
-
-        if (row.__depth === 0) {
-          const rowIdx = row.__index;
-          const stableKey = getStableRowKeyForRowIdx(rowIdx);
-
-          return (
-            <RowExpander
-              rowId={rowId}
-              isExpanded={expandedRows.has(stableKey)}
-              onCellExpand={() => {
-                setExpandedRows((er) => {
-                  if (er.has(stableKey)) {
-                    er.delete(stableKey);
-                  } else {
-                    er.add(stableKey);
-                  }
-                  return new Set(er);
-                });
-              }}
-            />
-          );
-        }
-
-        const expandedRecords = nestedRows[row.__index]?.final ?? [];
-        if (!expandedRecords.length) {
-          return (
-            <div className={styles.noDataNested}>
-              <Trans i18nKey="grafana-ui.table.nested-table.no-data">No data</Trans>
-            </div>
-          );
-        }
-
-        const nestedColumns = nestedColumnsMatrix[row.__index].columns;
-
-        return (
-          <div id={rowId}>
-            <DataGrid<TableRow, TableSummaryRow>
-              {...commonDataGridProps}
-              className={clsx(styles.grid, styles.gridNested)}
-              headerRowClass={clsx(styles.headerRow, hasNestedHeaders ? '' : styles.displayNone)}
-              headerRowHeight={hasNestedHeaders ? nestedHeaderHeightPx : 0}
-              bottomSummaryRows={hasNestedFooter ? [{}] : undefined}
-              summaryRowHeight={nestedFooterHeightPx}
-              onColumnResize={nestedResizeHandler}
-              columns={nestedColumns}
-              rows={expandedRecords}
-              renderers={{ ...renderers, noRowsFallback: <EmptyTablePlaceholder noValue={noValue} /> }}
-              onCellClick={onCellClick}
-              columnWidths={nestedColWidths}
-              onColumnWidthsChange={handleNestedColumnWidthsChange}
-            />
-          </div>
-        );
-      },
-      renderHeaderCell(props) {
-        return <div className="sr-only">{props.column.name}</div>;
-      },
-      width: COLUMN.EXPANDER_WIDTH,
-      minWidth: COLUMN.EXPANDER_WIDTH,
-    }),
-    [
-      styles.cellNested,
-      styles.grid,
-      styles.gridNested,
-      styles.headerRow,
-      styles.displayNone,
-      styles.noDataNested,
-      data.fields.length,
-      commonDataGridProps,
-      expandedRows,
-      getStableRowKeyForRowIdx,
-      nestedRows,
-      noValue,
-      onCellClick,
-      uniqueId,
-      nestedColWidths,
-      nestedResizeHandler,
-      handleNestedColumnWidthsChange,
-    ]
-  );
-
-  const fromFields = useCallback(
-    (
-      f: Field[],
-      widths: number[],
-      frame: DataFrame,
-      rawRows: TableRow[],
-      visibleRows: TableRow[]
-    ): FromFieldsResult => {
-      const result: FromFieldsResult = {
-        columns: [],
-        cellRootRenderers: {},
-      };
-
-      // Derive footer config from the fields being processed so nested tables use
-      // their own footer configuration rather than the top-level table's.
-      const fieldFooters: Array<TableFooterOptions | undefined> = [];
-      let isFieldUniformFooter = true;
-      let firstFooterReducers: string[] | undefined;
-      for (const field of f) {
-        const footer = field.config?.custom?.footer;
-        const reducers: string[] | undefined = footer?.reducers;
-
-        fieldFooters.push(footer);
-
-        // if reducers are undefined or empty on the footer, don't retain them for comparison.
-        if (reducers === undefined || reducers.length === 0) {
-          continue;
-        }
-
-        // first time we encounter a viable footer config, store it and move on.
-        if (firstFooterReducers === undefined) {
-          firstFooterReducers = reducers;
-          continue;
-        }
-
-        // for all other viable footer configs, check to see if the reducers match the first one we encountered.
-        if (
-          reducers.length !== firstFooterReducers.length ||
-          reducers.some((r, idx) => firstFooterReducers?.[idx] !== r)
-        ) {
-          isFieldUniformFooter = false;
-          break;
-        }
-      }
-
-      // Reuse pre-computed filter results — no re-scanning of rows needed.
-      // Top-level tables use filterResult from useFilteredRows; nested tables use the
-      // filterResult stored on their NestedRowEntry by useNestedRows.
-      const parentIndex = visibleRows[0]?.__parentIndex;
-      const { crossFilterRows, crossFilterTailRows } =
-        parentIndex == null ? filterResult : nestedRows[parentIndex].filterResult;
-
-      let lastRowIdx = -1;
-      // shared when whole row will be styled by a single cell's color
-      let rowCellStyle: Partial<CSSProperties> = {
-        color: undefined,
-        background: undefined,
-      };
-
-      for (let i = 0; i < f.length; i++) {
-        let field = f[i];
-        const cellOptions = getCellOptions(field);
-        const cellType = cellOptions.type;
-
-        // make sure we use mappings exclusively if they exist, ignore default thresholds mode
-        // we hack this by using the single color mode calculator
-        if (cellType === TableCellDisplayMode.Pill && (field.config.mappings?.length ?? 0 > 0)) {
-          field = {
-            ...field,
-            config: {
-              ...field.config,
-              color: {
-                ...field.config.color,
-                mode: FieldColorModeId.Fixed,
-                fixedColor: field.config.color?.fixedColor ?? FALLBACK_COLOR,
-              },
-            },
-          };
-          field.display = getDisplayProcessor({ field, theme });
-        }
-
-        // attach JSONCell custom display function to JSONView cell type
-        if (cellType === TableCellDisplayMode.JSONView || field.type === FieldType.other) {
-          field.display = displayJsonValue(field);
-        }
-
-        // For some cells, "aligning" the cell will mean aligning the inline contents of the cell with
-        // the text-align css property, and for others, we'll use justify-content to align the cell
-        // contents with flexbox. We always just get both and provide both when styling the cell.
-        const textAlign = getAlignment(field);
-        const justifyContent = getJustifyContent(textAlign);
-        const displayName = getDisplayName(field);
-        const headerCellClass = getHeaderCellStyles(theme, justifyContent);
-        const CellType = getCellRenderer(field, cellOptions);
-
-        const cellInspect = isCellInspectEnabled(field);
-        const showFilters = Boolean(field.config.filterable && onCellFilterAdded != null);
-        const showActions = cellInspect || showFilters;
-        const width = widths[i];
-
-        // helps us avoid string cx and emotion per-cell
-        const cellActionClassName = showActions
-          ? clsx('table-cell-actions', getCellActionStyles(theme, textAlign))
-          : undefined;
-
-        const shouldOverflow =
-          !IS_SAFARI_26 && typeof rowHeight !== 'string' && (shouldTextOverflow(field) || Boolean(maxRowHeight));
-        const textWrap = typeof rowHeight === 'string' || shouldTextWrap(field);
-        const canBeColorized = canFieldBeColorized(cellType, applyToRowBgFn);
-        const cellStyleOptions: TableCellStyleOptions = {
-          textAlign,
-          textWrap,
-          shouldOverflow,
-          maxHeight: maxRowHeight,
-        };
-
-        const defaultCellStyles = getDefaultCellStyles(theme, cellStyleOptions);
-        const cellSpecificStyles = getCellSpecificStyles(cellType, field, theme, cellStyleOptions);
-        const linkStyles = getLinkStyles(theme, canBeColorized);
-        const cellParentStyles = clsx(defaultCellStyles, linkStyles);
-        const maxHeightClassName = maxRowHeight ? getMaxHeightCellStyles(theme, cellStyleOptions) : undefined;
-        const styleFieldValue = field.config.custom?.styleField;
-        const styleField = styleFieldValue ? frame.fields.find(predicateByName(styleFieldValue)) : undefined;
-        const styleFieldName = styleField ? getDisplayName(styleField) : undefined;
-        const hasValidStyleField = Boolean(styleFieldName);
-
-        // TODO: in future extend this to ensure a non-classic color scheme is set with AutoCell
-
-        // this fires first
-        const renderCellRoot = (key: Key, props: CellRendererProps<TableRow, TableSummaryRow>): ReactNode => {
-          const rowIdx = props.row.__index;
-
-          // meh, this should be cached by the renderRow() call?
-          if (rowIdx !== lastRowIdx) {
-            lastRowIdx = rowIdx;
-
-            rowCellStyle.color = undefined;
-            rowCellStyle.background = undefined;
-
-            // generate shared styles for whole row
-            if (applyToRowBgFn != null) {
-              rowCellStyle = { ...rowCellStyle, ...applyToRowBgFn(rowIdx) };
-            }
-          }
-
-          let style: CSSProperties = { ...rowCellStyle };
-          if (canBeColorized) {
-            const value = props.row[props.column.key];
-            const displayValue = field.display!(value); // this fires here to get colors, then again to get rendered value?
-            const cellColorStyles = getCellColorInlineStyles(cellOptions, displayValue, applyToRowBgFn != null);
-            Object.assign(style, cellColorStyles);
-          }
-          if (hasValidStyleField) {
-            style = { ...style, ...parseStyleJson(props.row[styleFieldName!]) };
-          }
-
-          return (
-            <Cell
-              key={key}
-              {...props}
-              className={clsx(
-                props.className,
-                cellParentStyles,
-                cellSpecificStyles != null && maxRowHeight == null ? cellSpecificStyles : ''
-              )}
-              style={style}
-            />
-          );
-        };
-
-        result.cellRootRenderers[displayName] = renderCellRoot;
-
-        const renderBasicCellContent = (props: RenderCellProps<TableRow, TableSummaryRow>): JSX.Element => {
-          const rowIdx = props.row.__index;
-          const value = props.row[props.column.key];
-          // TODO: it would be nice to get rid of passing height down as a prop. but this value
-          // is cached so the cost of calling for every cell is low.
-          // NOTE: some cell types still require a height to be passed down, so that's why string-based
-          // cell types are going to just pass down the max cell height as a numeric height for those cells.
-          const height = rowHeightFn(props.row);
-
-          let result = (
-            <>
-              <CellType
-                cellOptions={cellOptions}
-                frame={frame}
-                field={field}
-                height={height}
-                rowIdx={rowIdx}
-                theme={theme}
-                value={value}
-                width={width}
-                timeRange={timeRange}
-                cellInspect={cellInspect}
-                showFilters={showFilters}
-                getActions={getCellActions}
-                disableSanitizeHtml={disableSanitizeHtml}
-                getTextColorForBackground={getTextColorForBackground}
-              />
-              {showActions && (
-                <TableCellActions
-                  field={field}
-                  value={value}
-                  displayName={displayName}
-                  cellInspect={cellInspect}
-                  showFilters={showFilters}
-                  className={cellActionClassName}
-                  setInspectCell={setInspectCell}
-                  onCellFilterAdded={onCellFilterAdded}
-                />
-              )}
-            </>
-          );
-
-          if (maxRowHeight != null) {
-            result = <div className={clsx(maxHeightClassName, cellSpecificStyles)}>{result}</div>;
-          }
-
-          return result;
-        };
-
-        // renderCellContent fires second.
-        let renderCellContent = renderBasicCellContent;
-
-        const tooltipFieldName = field.config.custom?.tooltip?.field;
-        if (tooltipFieldName) {
-          const tooltipField = frame.fields.find(predicateByName(tooltipFieldName));
-          if (tooltipField) {
-            const tooltipDisplayName = getDisplayName(tooltipField);
-            const tooltipCellOptions = getCellOptions(tooltipField);
-            const tooltipFieldRenderer = getCellRenderer(tooltipField, tooltipCellOptions);
-
-            const tooltipCellStyleOptions = {
-              textAlign: getAlignment(tooltipField),
-              // tooltips are free-floating overlays that should reveal the full value, so we
-              // always wrap their content and never inherit the per-row cell-height clamp
-              // (which would line-clamp/cut off the content).
-              textWrap: true,
-              shouldOverflow: false,
-            } satisfies TableCellStyleOptions;
-            const tooltipCanBeColorized = canFieldBeColorized(tooltipCellOptions.type, applyToRowBgFn);
-            const tooltipDefaultStyles = getDefaultCellStyles(theme, tooltipCellStyleOptions);
-            const tooltipSpecificStyles = getCellSpecificStyles(
-              tooltipCellOptions.type,
-              tooltipField,
-              theme,
-              tooltipCellStyleOptions
-            );
-            const tooltipLinkStyles = getLinkStyles(theme, tooltipCanBeColorized);
-            const tooltipClasses = getTooltipStyles(theme, textAlign);
-
-            const placement = field.config.custom?.tooltip?.placement ?? TableCellTooltipPlacement.Auto;
-            const tooltipWidth =
-              placement === TableCellTooltipPlacement.Left || placement === TableCellTooltipPlacement.Right
-                ? tooltipField.config.custom?.width
-                : width;
-
-            const tooltipProps = {
-              cellOptions: tooltipCellOptions,
-              classes: tooltipClasses,
-              className: clsx(
-                tooltipClasses.tooltipContent,
-                tooltipDefaultStyles,
-                tooltipSpecificStyles,
-                tooltipLinkStyles
-              ),
-              data: frame,
-              disableSanitizeHtml,
-              field: tooltipField,
-              getActions: getCellActions,
-              getTextColorForBackground,
-              gridRef,
-              placement,
-              renderer: tooltipFieldRenderer,
-              theme,
-              width: tooltipWidth,
-            } satisfies Partial<React.ComponentProps<typeof TableCellTooltip>>;
-
-            renderCellContent = (props: RenderCellProps<TableRow, TableSummaryRow>): JSX.Element => {
-              // cached so we don't care about multiple calls.
-              const tooltipHeight = rowHeightFn(props.row);
-              let tooltipStyle: CSSProperties = { ...rowCellStyle };
-              if (tooltipCanBeColorized) {
-                const tooltipDisplayValue = tooltipField.display!(props.row[tooltipDisplayName]);
-                const tooltipCellColorStyles = getCellColorInlineStyles(
-                  tooltipCellOptions,
-                  tooltipDisplayValue,
-                  applyToRowBgFn != null
-                );
-                Object.assign(tooltipStyle, tooltipCellColorStyles);
-              }
-
-              return (
-                <TableCellTooltip
-                  {...tooltipProps}
-                  height={tooltipHeight}
-                  rowIdx={props.row.__index}
-                  style={tooltipStyle}
-                >
-                  {renderBasicCellContent(props)}
-                </TableCellTooltip>
-              );
-            };
-          }
-        }
-
-        result.columns.push({
-          field,
-          key: displayName,
-          name: displayName,
-          width,
-          headerCellClass,
-          frozen: Math.min(frozenColumns, numFrozenColsFullyInView) > i,
-          renderCell: renderCellContent,
-          renderHeaderCell: ({ column, sortDirection }) => (
-            <HeaderCell
-              column={column}
-              rows={rawRows}
-              field={field}
-              filter={filter}
-              setFilter={setFilter}
-              disableKeyboardEvents={disableKeyboardEvents}
-              direction={sortDirection}
-              showTypeIcons={showTypeIcons}
-              parentIndex={parentIndex}
-              crossFilterRows={crossFilterRows}
-              crossFilterTailRows={crossFilterTailRows}
-              selectFirstCell={() => {
-                gridRef.current?.selectCell({ rowIdx: 0, idx: 0 });
-              }}
-            />
-          ),
-          renderSummaryCell: () => (
-            <SummaryCell
-              rows={visibleRows}
-              footers={fieldFooters}
-              field={field}
-              colIdx={i}
-              textAlign={getSummaryCellTextAlign(textAlign, cellType)}
-              rowLabel={isFieldUniformFooter && i === 0}
-              hideLabel={isFieldUniformFooter && i !== 0}
-            />
-          ),
-        });
-      }
-
-      return result;
-    },
-    [
-      applyToRowBgFn,
-      disableKeyboardEvents,
-      disableSanitizeHtml,
-      filter,
-      filterResult,
-      frozenColumns,
-      getCellActions,
-      getCellColorInlineStyles,
-      getTextColorForBackground,
-      maxRowHeight,
-      nestedRows,
-      numFrozenColsFullyInView,
-      onCellFilterAdded,
-      rowHeight,
-      rowHeightFn,
-      setFilter,
-      showTypeIcons,
-      theme,
-      timeRange,
-    ]
-  );
-
-  const nestedColumnsMatrix = useMemo(() => {
-    const result: FromFieldsResult[] = [];
-    if (!hasNestedFrames) {
-      return result;
-    }
-    for (const row of rows) {
-      if (row.__depth > 0) {
-        const rowNestedFrame = nestedData![row.__index]!;
-        result.push(
-          fromFields(
-            getVisibleFields(rowNestedFrame.fields),
-            nestedFieldWidths,
-            rowNestedFrame,
-            nestedRows[row.__index].raw,
-            nestedRows[row.__index].final
-          )
-        );
-      }
-    }
-    return result;
-  }, [rows, hasNestedFrames, nestedData, nestedRows, nestedFieldWidths, fromFields]);
-
-  const { columns, cellRootRenderers } = useMemo(() => {
-    const result = fromFields(visibleFields, widths, data, rows, sortedRows);
-
-    // if nested frames are present, augment the columns to include the nested table expander column.
-    if (!firstRowNestedData) {
-      return result;
-    }
-
-    // pre-calculate renderRow and expandedColumns based on the first nested frame's fields.
-    const renderRow = renderRowFactory(
-      firstRowNestedData.fields,
-      panelContext,
-      expandedRows,
-      enableSharedCrosshair,
-      getStableRowKeyForRowIdx
-    );
-
-    const expanderCellRenderer: CellRootRenderer = (key, props) => <Cell key={key} {...props} />;
-    result.cellRootRenderers[EXPANDED_COLUMN_KEY] = expanderCellRenderer;
-
-    // If we have nested frames, we need to add a column for the row expansion
-    result.columns.unshift(
-      buildNestedTableExpanderColumn(
-        nestedColumnsMatrix,
-        hasNestedHeaders,
-        nestedHeaderHeight,
-        nestedHasFooter,
-        nestedFooterHeight,
-        {
-          renderRow,
-          renderCell: (key, props) =>
-            nestedColumnsMatrix[props.row.__parentIndex!].cellRootRenderers[props.column.key](key, props),
-        }
-      )
-    );
-
-    return result;
-  }, [
-    buildNestedTableExpanderColumn,
-    data,
-    enableSharedCrosshair,
-    expandedRows,
-    firstRowNestedData,
-    fromFields,
-    getStableRowKeyForRowIdx,
-    hasNestedHeaders,
-    nestedColumnsMatrix,
-    nestedFooterHeight,
-    nestedHasFooter,
-    nestedHeaderHeight,
-    panelContext,
-    rows,
-    sortedRows,
-    visibleFields,
-    widths,
-  ]);
-
-  // invalidate columns on every structureRev change. this supports width editing in the fieldConfig.
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  const structureRevColumns = useMemo(() => columns, [columns, structureRev]);
-  const renderCellRoot: CellRootRenderer = useCallback(
-    (key, props) => cellRootRenderers[props.column.key](key, props),
-    [cellRootRenderers]
-  );
-
-  // we need to have variables with these exact names for the localization to work properly
-  const itemsRangeStart = pageRangeStart;
-  const displayedEnd = pageRangeEnd;
-
-  let rendered = (
-    <>
-      <DataGrid<TableRow, TableSummaryRow, string>
-        {...commonDataGridProps}
-        role={hasNestedFrames ? 'treegrid' : 'grid'}
-        ref={gridRef}
-        className={styles.grid}
-        columns={structureRevColumns}
-        rows={paginatedRows}
-        rowKeyGetter={rowKeyGetter}
-        isRowSelectionDisabled={() => initialRowIndex !== undefined}
-        selectedRows={selectedRows}
-        onSelectedRowsChange={setSelectedRows}
-        headerRowClass={clsx(styles.headerRow, noHeader ? styles.displayNone : '')}
-        headerRowHeight={headerHeight}
-        columnWidths={resetColumnWidths}
-        onColumnWidthsChange={resetColumnWidths != null ? () => {} : undefined}
-        onColumnResize={resizeHandler}
-        onCellClick={onCellClick}
-        onCellKeyDown={({ column, row }, event) => {
-          // if top-left cell, use default browser tabbing
-          if (column.key === columns[0].key && row.__index === 0 && event.shiftKey && event.key === 'Tab') {
-            event.preventGridDefault();
-            gridRef.current?.selectCell({ rowIdx: -1, idx: columns.length - 1 }); // select the far right cell of the header
-            return;
-          }
-
-          if (
-            disableKeyboardEvents ||
-            (hasNestedFrames && event.isDefaultPrevented()) // skip parent grid keyboard navigation if nested grid handled it
-          ) {
-            event.preventGridDefault();
-          }
-        }}
-        renderers={{
-          renderRow,
-          renderCell: renderCellRoot,
-          noRowsFallback: <EmptyTablePlaceholder noValue={noValue} />,
-        }}
-      />
-
-      {enablePagination && numRows > 0 && (
-        <div className={styles.paginationContainer}>
-          <Pagination
-            className="table-ng-pagination"
-            currentPage={page + 1}
-            numberOfPages={numPages}
-            showSmallVersion={smallPagination}
-            onNavigate={(toPage) => {
-              setPage(toPage - 1);
-            }}
-          />
-          {!smallPagination && (
-            <div className={styles.paginationSummary}>
-              {/* TODO: once TableRT is deprecated, we can update the localiziation
-                    string with the more consistent variable names */}
-              <Trans i18nKey="grafana-ui.table.pagination-summary">
-                {{ itemsRangeStart }} - {{ displayedEnd }} of {{ numRows }} rows
-              </Trans>
-            </div>
-          )}
-        </div>
-      )}
-
-      {tooltipState && (
-        <DataLinksActionsTooltip
-          links={tooltipState.links ?? []}
-          actions={tooltipState.actions}
-          coords={tooltipState.coords}
-          onTooltipClose={() => setTooltipState(undefined)}
-        />
-      )}
-
-      {inspectCell && (
-        <TableCellInspector
-          mode={inspectCell.mode ?? TableCellInspectorMode.text}
-          value={inspectCell.value}
-          onDismiss={() => setInspectCell(null)}
-        />
-      )}
-    </>
-  );
-
-  if (IS_SAFARI_26) {
-    rendered = <div className={styles.safariWrapper}>{rendered}</div>;
-  }
+  const rendered = IS_SAFARI_26 ? <div className={styles.safariWrapper}>{inner}</div> : inner;
 
   if (!tableHasGeoCell) {
     return rendered;
@@ -1148,52 +35,3 @@ export function TableNG(props: TableNGProps) {
     </Suspense>
   );
 }
-
-/**
- * this is passed to the top-level `renderRow` prop on DataGrid. applies aria attributes and custom event handlers.
- */
-const renderRowFactory =
-  (
-    fields: Field[],
-    panelContext: PanelContext,
-    expandedRows: Set<string>,
-    enableSharedCrosshair: boolean,
-    getStableKey: (rowIdx: number) => string
-  ) =>
-  // eslint-disable-next-line react/display-name
-  (key: React.Key, props: RenderRowProps<TableRow, TableSummaryRow>): React.ReactNode => {
-    const { row } = props;
-    const rowIdx = row.__index;
-    const isExpanded = expandedRows.has(getStableKey(rowIdx));
-
-    // Don't render non expanded child rows
-    if (row.__depth === 1) {
-      if (!isExpanded) {
-        return null;
-      }
-
-      // Add aria-expanded and aria-level to parent rows that have nested data
-      return <Row key={key} aria-level={row.__index + 1} aria-expanded={isExpanded} {...props} />;
-    }
-
-    const handlers: Partial<typeof props> = {};
-    if (enableSharedCrosshair) {
-      const timeField = fields.find((f) => f.type === FieldType.time);
-      if (timeField) {
-        handlers.onMouseEnter = () => {
-          panelContext.eventBus.publish(
-            new DataHoverEvent({
-              point: {
-                time: timeField?.values[rowIdx],
-              },
-            })
-          );
-        };
-        handlers.onMouseLeave = () => {
-          panelContext.eventBus.publish(new DataHoverClearEvent());
-        };
-      }
-    }
-
-    return <Row key={key} {...props} {...handlers} />;
-  };
